# swagger.yaml
openapi: 3.0.0
info:
  title: Ebooks Reader Backend API
  version: 1.0.1
  description: |
    Оновлена документація REST API для мобільного рідера книг. 
    - Додано multipart upload для книги та обкладинки (cover)
    - Додано підтримку Cloudinary для covers
    - Додано публічну бібліотеку книг (marketplace)
    - Додано поля isPublic, imageUrl
    - Оновлено всі схеми та приклади
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Bookmark:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        userId:
          type: integer
        position:
          type: string
        note:
          type: string
    Note:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        userId:
          type: integer
        content:
          type: string
        position:
          type: string
    ReadingProgress:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        userId:
          type: integer
        progress:
          type: number
          format: float
        position:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string
security:
  - bearerAuth: []
paths:
...
  /api/mobile/progress:
    get:
      summary: Отримання прогресу всіх книг користувача
      tags: [Mobile]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Список прогресу
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReadingProgress'
        401:
          description: Неавторизовано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/mobile/sync-progress:
    post:
      summary: Синхронізація прогресу читання
      tags: [Mobile]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookId:
                  type: integer
                progress:
                  type: number
                  format: float
                  minimum: 0
                  maximum: 1
                position:
                  type: string
      responses:
        200:
          description: Прогрес синхронізовано
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        400:
          description: Помилка валідації
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Неавторизовано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/mobile/bookmarks:
    get:
      summary: Отримання закладок
      tags: [Mobile]
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: query
          schema:
            type: integer
          description: ID книги (опціонально - якщо не вказано, повертаються всі закладки)
      responses:
        200:
          description: Список закладок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bookmark'
        400:
          description: Помилка валідації
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Неавторизовано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Створення закладки
      tags: [Mobile]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookId:
                  type: integer
                position:
                  type: string
                note:
                  type: string
      responses:
        201:
          description: Закладку створено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bookmark'
        400:
          description: Помилка валідації
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Неавторизовано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Книгу не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/mobile/bookmarks/{id}:
    put:
      summary: Редагування закладки
      tags: [Mobile]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                position:
                  type: string
                note:
                  type: string
      responses:
        200:
          description: Закладку оновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bookmark'
        400:
          description: Помилка валідації
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Неавторизовано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Закладку не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Видалення закладки
      tags: [Mobile]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: Закладку видалено
        404:
          description: Закладку не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Неавторизовано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/mobile/notes:
    get:
      summary: Отримання нотаток
      tags: [Mobile]
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: query
          schema:
            type: integer
          description: ID книги (опціонально - якщо не вказано, повертаються всі нотатки)
      responses:
        200:
          description: Список нотаток
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
        400:
          description: Помилка валідації
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Неавторизовано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Створення нотатки
      tags: [Mobile]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookId:
                  type: integer
                content:
                  type: string
                position:
                  type: string
      responses:
        201:
          description: Нотатку створено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        400:
          description: Помилка валідації
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Неавторизовано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Книгу не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/mobile/notes/{id}:
    put:
      summary: Редагування нотатки
      tags: [Mobile]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                position:
                  type: string
      responses:
        200:
          description: Нотатку оновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        400:
          description: Помилка валідації
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Неавторизовано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Нотатку не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Видалення нотатки
      tags: [Mobile]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: Нотатку видалено
        404:
          description: Нотатку не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Неавторизовано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/mobile/sync-notes:
    post:
      summary: Синхронізація нотаток
      tags: [Mobile]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookId:
                  type: integer
                notes:
                  type: array
                  items:
                    type: object
                    properties:
                      content:
                        type: string
                      position:
                        type: string
      responses:
        200:
          description: Нотатки синхронізовано
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
        400:
          description: Помилка валідації
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Неавторизовано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Книгу не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...